Добавлен прогресс бар при загруке данных из БД
